Introduction to Project
As it stands, I just finished a Syngnathidae (Seahorse, Pipefish, and Seadragon)-checker based on amino acid sequence. The amino acids will now be checked for recombinants (unnatural sequences, as two versions of the same allele were both split into alpha and beta sections, and four permutations created), then evaluated for if they're under directional (or positive) selection by at least 2/3 models. Z-scales, or physiochemical properties, are then obtained from each protein. The Z-scale protein data is then fed to DAPC. After dividing the proteins into k number of clusters, testing k's from 1-20, A-score spline interpolation was used to analyze the best k-value to use by finding the minimum number of clusters when intra-cluster variation goes down (called the "elbow"), like on a spectrum from one individual/group to everyone in a single group (one individual/group has, of course, the least amount of intra-group variation, but at some point, there is still relative uniformity within several (k) groups of numerous individuals. That k is used as the estimated number of groups that sampled individuals fall into). Once we know the number of groups there are among the sampled individuals, there becomes the question of who is in what group. 

DAPC then tests random individuals against each other, amongst the-best-k-number of groups. This happens 250 times. Individuals with strong similarities in their protein data, producing a high correlation between them, are likely in the same group (cluster). A matrix analysis then evaluates these correlations, and groups highly-correlated individuals with each other. This should ultimately produce the-best-k-number of groups (they need not be of the same sizes, either!). Each cluster now consists of individuals with similar MHII immune binding-site proteins, AKA supertypes. 

Also, 16S microbe half uses the taxonomizr code, but since taxonomizr doesn't work with these sequences well, and rentrez does, I'll reconstruct it. Current goal is to see if pident (or equivalent of 16S RNA match with the top result) is over 97. If yes, then the 16S RNA sequence is kept. Then phred will be checked to see if it's at least 25, followed by a checking of 99.9% or greater sequence identity to the top result. 16S RNA sequences which pass all of these checks will rarefied to 8158 reads and then their OTUs will be accessed. From this, the taxonomic order of the top result will be used. 

The microbial orders are then compared to MHII supertypes via the Wald's Z test. Higher Wald's Z scores between a microbial order and supertype indicate that they probably occurred together not by chance, but via selection pressures. In other words, the supertype was selected for in Syngnathidae in an environment with microbes of that order, due to it being an immune defense against microbes of that order.

These paragraphs are partially myself thinking aloud, and I'll refine the methodology as updates come. All methodology is based on Microbial-Host Dynamics And The Evolution Of Nearshore Diversity In Hippocampus Abdominalis (Ashe, 2021).